requirements.txt
Flask
gunicorn
datetime
mysql.connector
typing
sqlalchemy
urllib
logging
re
thread
threading
time
sys
disconnect
json
requests
cryptocode
pprint